Agrego la designacion de empleados segun tiempo libre, y fin e inicio de los turnos

+++ TP5/TP 5 v2/Pedido.cs
@@ -8,11 +8,11 @@ namespace TP_5_v2
 {
     class Pedido
     {
-        public float id;
+        public double id;
         public string tipoPedido;
 
 
-        public Pedido(float numeroPedido, string tipoPedido)
+        public Pedido(double numeroPedido, string tipoPedido)
         {
             this.id = numeroPedido;
             this.tipoPedido = tipoPedido;
